kernel: extras: base implementation of avc log spoofing
this exposes a new handler int ksu_handle_slow_avc_audit(u32 *tsid)
which will check if su_sid is going to be printed on the audit log.

Usage:
        ksu_handle_slow_avc_audit(&tsid);

on slow_avc_audit() on security/selinux/avc.c
This way, we replace sid right before that struct is created.

This can also be implemented in kprobes which will be on enxt commit.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>
[skye-pa1n: Minimal adaptation to RKSU]